Add null check for player's game session before deletion
Added a conditional check to ensure that `player.GameSession` is not null before attempting to delete the current game session from the repository. This prevents potential null reference exceptions by ensuring the deletion operation is only performed if there is an active game session.

+++ BananaGame.Application/Features/GameSession/Handlers/EndGameSessionCommandHandler.cs
@@ -40,7 +40,11 @@ namespace BananaGame.Application.Features.GameSession.Handlers
             player.UpdatePlayerStats(farthestLevel, totalLevelsPlayed, newTotalTimePlayed, newHihghestScore);
 
             await _playerRepository.UpdateAsync(player);
-            await _gameSessionRepository.DeleteAsync(currentSession!);
+
+            if (player.GameSession != null)
+            {
+                await _gameSessionRepository.DeleteAsync(currentSession!);
+            }
 
             await _unitOfWork.SaveChangesAsync(cancellationToken);
